requirements.txt
MLStructFP
numpy<2
opencv-python
Pillow
matplotlib
